+++ package/src/firestoreAdmin/index.js
@@ -4,7 +4,13 @@
 * Context:
 *   From tests
 */
-import {dbAdmin, /*eventually,*/ preheat_EXP} from "./dbAdmin.js"
+import { strict as assert } from 'assert'
+
+import {dbAdmin, preheat_EXP} from "./dbAdmin.js"
+
+import { afterAll } from '@jest/globals'
+
+const autoUnsubs = new Set();    // Set of () => ()
 
 /*
 * Note: Just doing 'const { collection, doc } = dbAdmin' does not work.
@@ -13,14 +19,62 @@ function collection(path) {
   return dbAdmin.collection(path);
 }
 
-function doc(path) {
-  return dbAdmin.doc(path);
+/*
+* Like Firebase Admin SDK's 'doc'.
+*
+* By restricting the API surface, we keep version updates or incompatibilities from leaking to test code.
+* Also, this allows us to fix some known problems.
+*/
+function doc(docPath) {   // (string) => DocumentReference like
+  const ref = dbAdmin.doc(docPath);
+
+  assert(ref.set && ref.get && ref.onSnapshot);
+
+  // Wrap us into the 'unsub' chain, so we can release resources if they are still listened to, once Jest times out.
+  //
+  function onSnapshot(handler) {    // ((...) => ...) => () => Promise of ()   // returned function is 'unsub'
+    let unsub = ref.onSnapshot(handler);
+
+    autoUnsubs.add(unsub);        // tbd. works?
+
+    return async () => {    // wrapper around 'unsub'
+      if (!unsub) fail("'unsub' called twice")
+
+      assert(autoUnsubs.has(unsub));
+      autoUnsubs.delete(unsub);
+      await unsub();
+      unsub = null;
+    };
+  }
+
+  return {
+    get: () => ref.get(),     // () => ...
+    set: (v) => ref.set(v),   // (...) => Promise of ...
+    onSnapshot
+  }
 }
+
+/*
+* Close any remaining listeners, when Jest is done.
+*
+* This counteracts what looks like a Firebase Admin SDK bug (incompatibility with Jest), where a remaining listener
+* causes Jest not to be able to return to OS prompt.
+*
+* Reference:
+*   "Jest does not exit tests cleanly with Firebase Firestore, an older version does. [...]" (Jest issues)
+*     -> https://github.com/facebook/jest/issues/11464
+*/
+afterAll( async () => {
+
+  //console.log(`!!! ${ autoUnsubs.size } listeners still running (cleaning them up)`);   // DEBUG
+
+  const proms = Array.from(autoUnsubs) .map( unsub => unsub() );   // Array of Promise of ()
+  await Promise.all(proms);
+})
 
 export {
   collection,
   doc,
-  //eventually,
   //
   preheat_EXP
 }